// web/src/openapi/system/dasd.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * API description of the DASD system
 */
export interface System {
  /**
   * DASD devices
   */
  devices: Device[];
}
export interface Device {
  channel: string;
  deviceName: string;
  type: string;
  diag: boolean;
  accessType: string;
  partitionInfo: string;
  status: string;
  active: boolean;
  formatted: boolean;
}